Making some changes to material initializations
Added an init_materials routine which sets up moderator and
fuel materials and adds them to a materailList vector.  We
can either have accessor functions that actually use the materials,
or just have a get_ptr routine.  The latter is probably better, but
that decision can be made when we start doing the particle tracking.