Added fissionNeutron and simulate() functions
simulate() will simulate the entire life of a neutron until it
causes fission, leaks from the system, or gets absorbed.  It returns
an integer.  I think a positive value should indicate the number of
fissions caused, a negative value should represent the negative of
the surface id across which the neutron leaked, and a 0 represents
absorption of the neutron.  These will be useful for performing tallies.

fissionNeutron() takes a neutron as an argument and returns a new
particle.  It is called after a neutron is simulated to produce fission
neutrons from the simulated particle.  The particle's position and energy
can be used to generate the neutrons.

Both of these routines are currently empty.